Add: mql-tooling package for highlighting+linting MQL
I was going to add this all to mql-worker, but I saw that if I did that, it would add a number of higher-level dependencies to that package. So I decided to add another package instead. This will be the base that we build additional features off of, such as GoTo Definition.
