+++ packages/default-parts/package.json
@@ -26,16 +26,11 @@
   "dependencies": {
     "@jupyterlab/codeeditor": "~1.1.0",
     "@jupyterlab/codemirror": "~1.1.3",
-    "@mavenomics/bindings": "^0.1.0",
     "@mavenomics/coreutils": "^0.1.0",
-    "@mavenomics/mql-worker": "^0.1.0",
     "@mavenomics/parts": "^0.1.0",
     "@mavenomics/table": "^0.1.0",
     "@phosphor/application": "^1.6.0",
     "@phosphor/widgets": "^1.6.0",
-    "@types/codemirror": "^0.0.76",
-    "codemirror": "~5.47.0",
-    "jshint": "^2.10.2",
     "moment": "^2.24.0",
     "rc-calendar": "^9.10.10",
     "rc-select": "^9.0.0",
@@ -43,7 +38,6 @@
     "rc-tooltip": "^3.7.3"
   },
   "peerDependencies": {
-    "codemirror": ">=5.0.0",
     "react": ">=16.8.0"
   }
 }
