+++ packages/mql-tooling/.gitignore
@@ -1,0 +1,2 @@
+lib/
+mql.worker.js
