+++ packages/mql-tooling/README.md
@@ -1,0 +1,4 @@
+# `@mavenomics/mql-tooling`
+
+Language and editor tooling for the MQL engine, including highlighters and
+linters for CodeMirror in both the MQL dialect and the JSMQL hybrid dialect.
