+++ packages/mql-tooling/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@mavenomics/mql-tooling",
+  "version": "0.1.0",
+  "description": "Language tooling for MQL, including CodeMirror highlighting and linting",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "repository": "https://github.com/Mavenomics/MavenWorks",
+  "author": "Mavenomics, Inc",
+  "license": "GPL-3.0",
+  "files": [
+    "lib/",
+    "mql.worker.js"
+  ],
+  "scripts": {
+    "build": "tsc --build",
+    "build:ci": "echo pass",
+    "bundle:ci": "echo pass",
+    "clean": "rimraf ./lib *.tsbuildinfo",
+    "lint": "tslint",
+    "test": "echo No tests for @mavenomics/mql-tooling",
+    "docs": "echo pass"
+  },
+  "dependencies": {
+    "@mavenomics/bindings": "^0.1.0",
+    "@mavenomics/coreutils": "^0.1.0",
+    "@mavenomics/mql-worker": "^0.1.0",
+    "@mavenomics/table": "^0.1.0",
+    "@phosphor/application": "^1.6.0",
+    "@phosphor/coreutils": "^1.3.1",
+    "@phosphor/disposable": "1.3.0",
+    "@phosphor/widgets": "^1.9.3",
+    "codemirror": "~5.47.0",
+    "jshint": "^2.10.2"
+  },
+  "devDependencies": {
+    "@types/codemirror": "^0.0.79",
+    "rimraf": "^2.6.3",
+    "webpack": "^4.29.6"
+  }
+}
