+++ packages/mql-tooling/src/jshint.d.ts
