+++ packages/mql-tooling/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "extends": "../../base.tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./lib",
+    "rootDir": "./src",
+    // Temporarily disabling strict mode
+    "strict": false
+  },
+  "include": [
+    "./src/**/*"
+  ],
+  "references": [
+    { "path": "../bindings" },
+    { "path": "../coreutils" },
+    { "path": "../mql-worker" },
+    { "path": "../table" }
+  ]
+}
