+++ packages/mql-tooling/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../base.tslint.json"
+}
